set up for redux
