+++ BottomUp_UI/src/index.js
@@ -1,53 +1,15 @@
 import React, {Component} from 'react';
 import ReactDOM from 'react-dom';
-import MainLayout from './components/main_layout/main_layout'
-import ButtonList from './components/right-side/button_list'
-import Header from './components/header_bar/header';
+import { Provider } from 'react-redux';
+import { createStore, applyMiddleware } from 'redux';
 
-class App extends Component {
-    constructor(props) {
-        super(props);
+import reducers from './reducers';
+import App from './Components/app';
 
-        this.state = {
-            row: '0',
-            col: '0',
-            width: '0',
-            printCase: 4
-        };
-    }
+const createStoreWithMiddleware = applyMiddleware()(createStore);
 
-    onChangeState(row, col, width) {
-        this.setState({
-            row: row,
-            col: col,
-            width: width
-        });
-    }
-
-    render() {
-        const changeState = (row, col, width) => this.onChangeState(row, col, width);
-
-        return (
-            <div className="wrapper">
-                <div className="head">
-                    <Header onChangeState={changeState}/>
-                </div>
-                <div className="left">
-                    <MainLayout
-                        row={this.state.row}
-                        col={this.state.col}
-                        printCase={this.state.printCase}
-                    />
-                </div>
-                <div className="right">
-                    <ButtonList
-                        onPrintCaseSelect = {printCase => this.setState({printCase})}
-                    />
-                </div>
-            </div>
-
-        );
-    }
-}
-
-ReactDOM.render(<App/>, document.querySelector('.container'));
+ReactDOM.render(
+  <Provider store= { createStoreWithMiddleware(reducers) }>
+  <App/>
+  </Provider>
+  , document.querySelector('.container'));
