+++ BottomUp_UI/src/reducers/index.js
@@ -1,0 +1,7 @@
+import { combineReducers } from 'redux';
+
+const rootReducer = combineReducers({
+
+});
+
+export default rootReducer;
